% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_applying.R
\name{apply_overlay}
\alias{apply_overlay}
\title{Apply Selected Overlay}
\usage{
apply_overlay(studbook = NULL, overlay_to_use = NULL, verbose = TRUE,
  remove = TRUE, add_IsHypothetical = TRUE)
}
\arguments{
\item{studbook}{Studbook name}

\item{overlay_to_use}{Character name of which overlay to use}

\item{verbose}{Logical of whether or not to print details}

\item{remove}{Logical of whether or not to remove the (applied) overlay
components}

\item{add_IsHypothetical}{Logical of whether or not to add "IsHypothetical"
with a value of 0 in instances where the overlay is not applied}
}
\value{
Studbook as a named listed of database tables with overlay applied
}
\description{
Apply Selected Overlay
}
\examples{


}
