% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{apply_overlay_master}
\alias{apply_overlay_master}
\title{Apply the master table overlay}
\usage{
apply_overlay_master(studbook = studbook, overlay_UID_to_apply = NULL,
  add_IsHypothetical = TRUE)
}
\arguments{
\item{studbook}{Studbook name}

\item{overlay_UID_to_apply}{ID of the overlay to apply}

\item{add_IsHypothetical}{Logical of whether or not to add "IsHypothetical"
with a value of 0 in instances where the overlay is not applied}
}
\value{
Studbook master table with overlay applied
}
\description{
Apply the master table overlay
}
\examples{

}
