% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_fields.R
\name{apply_UDF}
\alias{apply_UDF}
\title{Apply Selected UDF (and Optionally Restrict Individuals)}
\usage{
apply_UDF(studbook = NULL, udf = NULL, levels = NULL,
  retain = TRUE, sb_ids = NULL)
}
\arguments{
\item{studbook}{Studbook name}

\item{udf}{Name of UDF to apply}

\item{levels}{Vector of levels of the UDF}

\item{retain}{Logical of whether levels should be kept (TRUE) or 
dropped (FALS)}

\item{sb_ids}{Character vector of studbook IDs to further limit the output}
}
\value{
Studbook IDs of all individuals that match the levels of the UDF
 as requested, with additional optional restrictions
}
\description{
Apply Selected UDF (and Optionally Restrict Individuals)
}
