% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_window_functions.R
\name{check_dates}
\alias{check_dates}
\title{Apply Date Window}
\usage{
check_dates(dates = NULL, date_window = NULL)
}
\arguments{
\item{dates}{Character vector of locations to be tested}

\item{date_window}{Character vector defining the dates of interest}
}
\value{
Logical vector of whether each date in input vector is one 
 of interest
}
\description{
Apply Date Window
}
\examples{

}
