% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institutional_windows.R
\name{check_institutions}
\alias{check_institutions}
\title{Apply Institutional Window}
\usage{
check_institutions(locations = NULL, location_window = NULL)
}
\arguments{
\item{locations}{Character vector of locations to be tested}

\item{location_window}{Character vector defining the locations of interest}
}
\value{
Logical vector of whether each location is input vector is one 
 of interest
}
\description{
Apply Institutional Window
}
