% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_window_functions.R
\name{create_date_window}
\alias{create_date_window}
\title{Create Date Window}
\usage{
create_date_window(starting_date = NULL, ending_date = NULL,
  starting_buffer = 0, ending_buffer = 0, buffer_units = "days")
}
\arguments{
\item{starting_date}{Character of the starting date (YYYY-MM-DD)}

\item{ending_date}{Character of the ending date (YYYY-MM-DD)}

\item{starting_buffer}{Amount of time to not include from the starting date}

\item{ending_buffer}{Amount of time not to include from the ending date}

\item{buffer_units}{Currently only supported for "days"}
}
\value{
Beginning and ending days of the date window.
}
\description{
Create Date Window
}
\examples{


}
