% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{events}
\alias{events}
\title{Determine relevant events for everyone}
\usage{
events(studbook = NULL, exclude = NULL, starting_date = NULL,
  ending_date = NULL, starting_buffer = 0, ending_buffer = 0,
  buffer_units = "days")
}
\arguments{
\item{studbook}{Studbook}

\item{exclude}{Character vector of StudbookIDs to exclude from the pedigree}

\item{starting_date}{Character of the starting date (YYYY-MM-DD)}

\item{ending_date}{Character of the ending date (YYYY-MM-DD)}

\item{starting_buffer}{Amount of time to not include from the starting date}

\item{ending_buffer}{Amount of time not to include from the ending date}

\item{buffer_units}{Currently only supported for "days"}
}
\value{
list of table of (non-transfer) events and table of transfers
}
\description{
Determine relevant events for everyone
}
