% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{examine_overlays}
\alias{examine_overlays}
\title{Determine how many overlays exist and parse the overlay tables}
\usage{
examine_overlays(studbook = NULL, verbose = FALSE, full_return = FALSE)
}
\arguments{
\item{studbook}{Studbook name}

\item{verbose}{Logical about whether or not to print}

\item{full_return}{Logical about whether to return the overlay or 
just details}
}
\value{
List of overlay tables, split by overlay.
}
\description{
Determine how many overlays exist and parse the overlay tables
}
\examples{

}
