% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined_fields.R
\name{examine_UDFs}
\alias{examine_UDFs}
\title{Determine how many UDFs exist, print, and return list}
\usage{
examine_UDFs(studbook = NULL, verbose = TRUE)
}
\arguments{
\item{studbook}{Studbook name}

\item{verbose}{Logical about whether or not to print the UDFs}
}
\value{
List of UDFs, each containing its levels.
}
\description{
Determine how many UDFs exist, print, and return list
}
