% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{inbreeding}
\alias{inbreeding}
\title{Calculate the inbreeding of a pedig_list object}
\usage{
inbreeding(pedig_list)
}
\arguments{
\item{pedig_list}{pedig_list object}
}
\value{
data frame with individual and parental ids and inbreeding values
}
\description{
Calculate the inbreeding of a pedig_list object
}
