% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{kinship.pedig_list}
\alias{kinship.pedig_list}
\title{Calculate the kinship of a pedig_list object}
\usage{
\method{kinship}{pedig_list}(id, ...)
}
\arguments{
\item{id}{pedig_list object}

\item{...}{optional arguments}
}
\value{
kinship matrix
}
\description{
Calculate the kinship of a pedig_list object
}
