% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studbook_loading.R
\name{load_poplink_studbook}
\alias{load_poplink_studbook}
\title{Load PopLink Studbook}
\usage{
load_poplink_studbook(db_name = NULL, db_tables = c("Master", "Event"),
  overlay = TRUE, udf = TRUE, verbose = FALSE)
}
\arguments{
\item{db_name}{Database name}

\item{db_tables}{Names of database tables to retrieve}

\item{overlay}{Logical of whether or not to retrieve the overlay}

\item{udf}{Logical of whether or not to retrieve the user defined field}

\item{verbose}{Logical of whether or not to print database and 
overlay details}
}
\value{
Studbook as a named listed of database tables
}
\description{
Load PopLink Studbook
}
