% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstud.R
\docType{package}
\name{Rstud}
\alias{Rstud}
\alias{Rstud-package}
\title{Process and analyze studbooks}
\description{
Provides utilities for importing, processing, editing, 
  analyzing, and exporting pedigree population databases (studbooks).
}
\keyword{package}
