% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_functions.R
\name{examine_udfs}
\alias{examine_udfs}
\title{Determine how many UDFs exist, print, and return list}
\usage{
examine_udfs(studbook = NULL, verbose = TRUE)
}
\arguments{
\item{studbook}{Studbook name}

\item{verbose}{Logical about whether or not to print the UDFs}
}
\value{
List of overlay tables, split by overlay.
}
\description{
Determine how many UDFs exist, print, and return list
}
\examples{

}
